+++ routes/web.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Route;
 use N3XT0R\LaravelWebdavServer\Http\Controllers\WebDavController;
 
 Route::match([
+    'OPTIONS',
     'GET',
     'PUT',
     'DELETE',